// dsl/packages/cli/test/test-specs/gen-out/client/typescript-client/src/model/PatchableScalarRecord_Substitution.ts
export type PatchableScalarRecord_Substitution = {
	readonly key: RSDString;
	readonly version: RSDString;
	readonly value: Range;
	readonly value_Null: Range | null;
	readonly value_Opt?: Range;
	readonly value_Opt_Null?: Range | null;
	readonly list: Range[];
	readonly list_Null: Range[] | null;
	readonly list_Opt?: Range[];
	readonly list_Opt_Null?: Range[] | null;
};
// ...
type $ListReplace = ListReplace<Range>;
type $ListMerge = ListMergeAddRemove<Range, Range>;
type $ListPatch = $ListReplace | $ListMerge;

type $List_NullReplace = ListReplace<Range>;
type $List_NullMerge = ListMergeAddRemove<Range, Range>;
type $List_NullPatch = $List_NullReplace | $List_NullMerge;

type $List_OptReplace = ListReplace<Range>;
type $List_OptMerge = ListMergeAddRemove<Range, Range>;
type $List_OptPatch = $List_OptReplace | $List_OptMerge;

type $List_Opt_NullReplace = ListReplace<Range>;
type $List_Opt_NullMerge = ListMergeAddRemove<Range, Range>;
type $List_Opt_NullPatch = $List_Opt_NullReplace | $List_Opt_NullMerge;

export type PatchableScalarRecord_SubstitutionPatch = {
	readonly key: RSDString;
	readonly version: RSDString;
	readonly value?: Range;
	readonly value_Null?: Range | null;
	readonly value_Opt?: Range | null;
	readonly value_Opt_Null?: Range | null;
	readonly list?: $ListPatch;
	readonly list_Null?: $List_NullPatch | null;
	readonly list_Opt?: $List_OptPatch | null;
	readonly list_Opt_Null?: $List_Opt_NullPatch | null;
};